tools/hello-world/pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Minimal hello-world tool for experiment template demonstration"
authors = [
    {name = "Experiment Template", email = "template@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
